+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib-java/src/main/resources/META-INF/azure-intellij-resource-connector-lib-java.xml
@@ -5,11 +5,11 @@
     <extensions defaultExtensionNs="com.intellij">
         <runConfigurationExtension implementation="com.microsoft.azure.toolkit.intellij.connector.dotazure.DotEnvRunConfigurationExtension"/>
         <annotator language="JAVA" implementationClass="com.microsoft.azure.toolkit.intellij.connector.annotator.ResourceConnectionAnnotator"/>
-        <typedHandler implementation="com.microsoft.azure.toolkit.intellij.connector.spring.PropertiesAutoPopupHandler" id="azPropertiesAuto"/>
-        <lookup.charFilter implementation="com.microsoft.azure.toolkit.intellij.connector.spring.PropertyCharFilter" id="azProperties"/>
-        <completion.contributor language="Properties" id="azPropertiesCompletion" order="before propertiesCompletion" implementationClass="com.microsoft.azure.toolkit.intellij.connector.spring.SpringPropertiesCompletionContributor"/>
+        <typedHandler implementation="com.microsoft.azure.toolkit.intellij.connector.spring.properties.SpringPropertiesAutoPopupHandler" id="azPropertiesAuto"/>
+        <lookup.charFilter implementation="com.microsoft.azure.toolkit.intellij.connector.spring.properties.SpringPropertyCharFilter" id="azProperties"/>
+        <completion.contributor language="Properties" id="azPropertiesCompletion" order="before propertiesCompletion" implementationClass="com.microsoft.azure.toolkit.intellij.connector.spring.properties.SpringPropertiesCompletionContributor"/>
         <completion.contributor language="JAVA" implementationClass="com.microsoft.azure.toolkit.intellij.connector.completion.ResourceConnectionCompletionContributor"/>
-        <codeInsight.lineMarkerProvider language="Properties" implementationClass="com.microsoft.azure.toolkit.intellij.connector.spring.SpringPropertiesLineMarkerProvider"/>
+        <codeInsight.lineMarkerProvider language="Properties" implementationClass="com.microsoft.azure.toolkit.intellij.connector.spring.properties.SpringPropertiesLineMarkerProvider"/>
         <codeInsight.lineMarkerProvider language="JAVA" implementationClass="com.microsoft.azure.toolkit.intellij.connector.marker.ResourceConnectionLineMarkerProvider"/>
     </extensions>
